move test data generation to separate file
